feat: creating utils functions to better structurate the project.
