+++ client/pages/index.tsx
@@ -1,93 +1,23 @@
 import type { NextPage } from "next";
-import Head from "next/head";
 
-import styles from "../styles/Home.module.css";
-import React, { useEffect, useState } from "react";
+import React from "react";
 import SignMessage from "../components/SignMessage";
-declare let window: any;
+import { useWallet } from "../hooks/useWallet";
+
 const Home: NextPage = () => {
-  const [currentAccount, setCurrentAccount] = useState<string>("");
+  const { currentAccount, connectWalletAction, disconnect } = useWallet();
 
-  const checkIfWalletIsConnected = async () => {
-    try {
-      const { ethereum } = window;
-
-      if (!ethereum) {
-        console.log("Make sure you have MetaMask!");
-        return;
-      } else {
-        console.log("We have the ethereum object", ethereum);
-
-        const accounts = await ethereum.request({ method: "eth_accounts" });
-
-        if (accounts.length !== 0) {
-          const account = accounts[0];
-          console.log("Found an authorized account:", account);
-          setCurrentAccount(account);
-        } else {
-          console.log("No authorized account found");
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-  const connectWalletAction = async () => {
-    try {
-      const { ethereum } = window;
-
-      if (!ethereum) {
-        alert("Get MetaMask!");
-        return;
-      }
-
-      const accounts = await ethereum.request({
-        method: "eth_requestAccounts",
-      });
-
-      console.log("Connected", accounts[0]);
-      setCurrentAccount(accounts[0]);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  useEffect(() => {
-    checkIfWalletIsConnected();
-  }, []);
-  useEffect(() => {
-    const checkNetwork = async () => {
-      try {
-        if (window.ethereum.networkVersion !== "3") {
-          alert("Please connect to Ropsten!");
-        }
-      } catch (error) {
-        console.log(error);
-      }
-    };
-
-    if (currentAccount) {
-      checkNetwork();
-    }
-  }, [currentAccount]);
-  const disconnect = () => {
-    setCurrentAccount("");
-  };
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <button
-        className="w-96 text-white bg-gray-800 hover:bg-gray-900 focus:outline-none focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-gray-800 dark:hover:bg-gray-700 dark:focus:ring-gray-700 dark:border-gray-700"
-        onClick={currentAccount ? disconnect : connectWalletAction}
-      >
-        {currentAccount ? "Disconnect" : "Connect Wallet To Get Started"}
-      </button>
-      {currentAccount && <SignMessage currentAccount={currentAccount} />}
+    <div>
+      <div className="w-full flex justify-center  mt-4">
+        <button
+          className="w-96 text-white bg-gray-800 hover:bg-gray-900 focus:outline-none focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-gray-800 dark:hover:bg-gray-700 dark:focus:ring-gray-700 dark:border-gray-700"
+          onClick={currentAccount ? disconnect : connectWalletAction}
+        >
+          {currentAccount ? "Disconnect" : "Connect Wallet To Get Started"}
+        </button>
+      </div>
+      {currentAccount && <SignMessage />}
     </div>
   );
 };
