+++ client/types/INFTInfo.ts
@@ -1,0 +1,4 @@
+export interface INFTInfo {
+  owner: string;
+  tokenUri: string;
+}
